flirt-bot/requirements.txt
telethon==1.29.2
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.8.5
asyncio==3.4.3
colorama==0.4.6
cryptg==0.4.0
openai==1.3.5
redis==5.0.1 